requirements.txt
airflow-providers-clickhouse==0.0.1
pandas==2.0.3
requests==2.31.0
pyspark==3.5.0
apache-airflow-providers-postgres==5.11.1
mimesis==6.1.1
kafka-python==2.2.15
clickhouse-driver==0.2.6
com.github.housepower:clickhouse-native-jdbc:2.6.5
apache-airflow-providers-apache-spark==4.3.0
minio==7.1.0
boto3==1.28.57
